Publish CaseCreatedEvent and correct events on state changes
https://eaflood.atlassian.net/browse/FGP-529

- removes old events and dead code
- switched to publishing CaseStatusUpdated with FQ statuses (hard-coded
  until workflows are updated to include lifecycle state machines)
- broke case integration tests appart to make it easier to manage
- db and queues get cleaned up before integration test runs
