+++ .env.example
@@ -12,8 +12,6 @@ AWS_ACCESS_KEY_ID=test
 AWS_SECRET_ACCESS_KEY=test
 AWS_ENDPOINT_URL=http://localstack:4566
 
-CREATE_NEW_CASE_SQS_URL=http://sqs.eu-west-2.127.0.0.1:4566/000000000000/create_new_case
-
 CW__SNS__CASE_CREATED_TOPIC_ARN=arn:aws:sns:eu-west-2:000000000000:cw__sns__case_created
 CW__SNS__CASE_STATUS_UPDATED_TOPIC_ARN=arn:aws:sns:eu-west-2:000000000000:cw__sns__case_status_updated
 CW__SQS__CREATE_NEW_CASE_URL=http://sqs.eu-west-2.127.0.0.1:4566/000000000000/cw__sqs__create_new_case
