+++ compose/start-localstack.sh
@@ -53,22 +53,27 @@ function subscribe_queue_to_topic() {
   awslocal sns subscribe --topic-arn $topic_arn --protocol sqs --notification-endpoint $queue_arn --attributes '{ "RawMessageDelivery": "true" }'
 }
 
-function setup_topic_and_queues() {
+function create_topic_and_queue() {
   local topic_name=$1
+  local queue_name=$2
 
   local topic_arn=$(create_topic $topic_name)
-  local queue_arn=$(create_queue $topic_name)
+  local queue_arn=$(create_queue $queue_name)
 
   subscribe_queue_to_topic $topic_arn $queue_arn
 }
 
-# ----- Commands -----
-create_queue "create_new_case"
-create_queue "cw__sqs__create_new_case"
-create_queue "cw__sqs__update_case_status"
 
+create_topic_and_queue "cw__sns__case_created" "cw__sqs__case_created" &
+create_topic_and_queue "cw__sns__case_status_updated" "gas__sqs__update_status" &
 
-# ----- Events ------
-setup_topic_and_queues "case_stage_updated"
-setup_topic_and_queues "cw__sns__case_created"
-setup_topic_and_queues "cw__sns__case_status_updated"
+create_topic_and_queue "agreement_status_updated" "gas__sqs__update_agreement_status" &
+create_topic_and_queue "gas__sns__grant_application_created" "gas__sqs__grant_application_created" &
+create_topic_and_queue "gas__sns__grant_application_status_updated" "gas__sqs__grant_application_status_updated" &
+create_topic_and_queue "gas__sns__create_new_case" "cw__sqs__create_new_case" &
+create_topic_and_queue "gas__sns__update_case_status" "cw__sqs__update_case_status" &
+create_topic_and_queue "gas__sns__create_agreement" "create_agreement" &
+
+wait
+
+echo "SNS/SQS ready"
