+++ package.json
@@ -12,16 +12,10 @@
     "test": "npm run test:unit -- --run && npm run test:integration -- --run",
     "test:unit": "vitest --config vitest.config.js --dir src",
     "test:integration": "vitest --config test/vitest.config.js --dir test",
-    "test:integration:report": "vitest --config test/vitest.config.js --dir test --reporter=html --run",
-    "test:integration:service": "vitest --config test/vitest.config.js test/integration/service --run",
-    "test:integration:events": "vitest --config test/vitest.config.js test/integration/events --run",
-    "test:integration:workflows": "vitest --config test/vitest.config.js test/integration/workflows --run",
-    "test:report": "npm run test:unit -- --run && npm run test:integration:report",
-    "test:report:custom": "node scripts/generate-test-report.js",
-    "dev": "node --env-file .env src/main.js",
+    "dev": "node --env-file .env --watch src/main.js",
     "publish:case:new": "node scripts/publish-create-new-case.js",
     "publish:case:pmf": "node scripts/publish-create-new-case.js pigs-might-fly",
-    "publish:case:agreement": "node scripts/publish-case-agreement-update-command.js",
+    "publish:case:status:update": "node scripts/publish-case-status.js",
     "postinstall": "npm run setup:husky && npm run setup:env",
     "setup:env": "cp --update=none .env.example .env || true",
     "setup:husky": "node -e \"try { (await import('husky')).default() } catch (e) { if (e.code !== 'ERR_MODULE_NOT_FOUND') throw e }\" --input-type module",
@@ -67,7 +61,6 @@
   },
   "devDependencies": {
     "@vitest/coverage-v8": "3.2.4",
-    "@vitest/ui": "3.2.4",
     "eslint": "9.29.0",
     "eslint-config-prettier": "10.1.5",
     "husky": "9.1.7",
