--- scripts/generate-test-report.js
@@ -1,374 +1,0 @@
-#!/usr/bin/env node
-
-import { execSync } from "node:child_process";
-import fs from "node:fs/promises";
-import path from "node:path";
-
-const REPORTS_DIR = "test/reports";
-const CUSTOM_REPORT_FILE = `${REPORTS_DIR}/comprehensive-report.html`;
-
-// Generate comprehensive test report
-async function generateComprehensiveReport() {
-  console.log("🔄 Generating comprehensive test report...");
-
-  try {
-    // Ensure reports directory exists
-    await fs.mkdir(REPORTS_DIR, { recursive: true });
-
-    // Run tests and capture output
-    let testOutput = "";
-    let testsPassed = 0;
-    let testsFailed = 0;
-    let totalTests = 0;
-
-    try {
-      // Run integration tests with JSON reporter
-      const result = execSync(
-        "npx vitest --config test/vitest.config.js --dir test --reporter=json --run",
-        {
-          encoding: "utf8",
-          timeout: 300000, // 5 minutes
-        },
-      );
-
-      const testResults = JSON.parse(result);
-      testsPassed = testResults.numPassedTests || 0;
-      testsFailed = testResults.numFailedTests || 0;
-      totalTests = testResults.numTotalTests || 0;
-
-      testOutput = formatTestResults(testResults);
-    } catch (error) {
-      console.log("⚠️  Tests had failures, capturing results...");
-
-      // Try to parse JSON from stdout if available
-      try {
-        const jsonMatch = error.stdout?.match(/\{[\s\S]*\}/);
-        if (jsonMatch) {
-          const testResults = JSON.parse(jsonMatch[0]);
-          testsPassed = testResults.numPassedTests || 0;
-          testsFailed = testResults.numFailedTests || 0;
-          totalTests = testResults.numTotalTests || 0;
-          testOutput = formatTestResults(testResults);
-        } else {
-          testOutput = error.stdout || error.message || "Test execution failed";
-        }
-      } catch (parseError) {
-        console.log("Could not parse test results JSON");
-        testOutput = error.stdout || error.message || "Test execution failed";
-      }
-    }
-
-    // Generate HTML report
-    const html = generateHTML({
-      timestamp: new Date().toISOString(),
-      testsPassed,
-      testsFailed,
-      totalTests,
-      testOutput,
-      passRate:
-        totalTests > 0 ? ((testsPassed / totalTests) * 100).toFixed(1) : 0,
-    });
-
-    // Write HTML report
-    await fs.writeFile(CUSTOM_REPORT_FILE, html, "utf8");
-
-    console.log("✅ Comprehensive test report generated!");
-    console.log(
-      `📊 Tests: ${testsPassed} passed, ${testsFailed} failed, ${totalTests} total`,
-    );
-    console.log(`📁 Report saved to: ${CUSTOM_REPORT_FILE}`);
-    console.log(
-      `🌐 Open in browser: file://${path.resolve(CUSTOM_REPORT_FILE)}`,
-    );
-  } catch (error) {
-    console.error("❌ Error generating report:", error.message);
-    process.exit(1);
-  }
-}
-
-function formatTestResults(testResults) {
-  if (!testResults || !testResults.testResults) {
-    return "No test results available";
-  }
-
-  let formattedOutput = `📊 TEST EXECUTION SUMMARY
-━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-
-Total Test Suites: ${testResults.numTotalTestSuites}
-✅ Passed Suites: ${testResults.numPassedTestSuites}
-❌ Failed Suites: ${testResults.numFailedTestSuites}
-
-Total Tests: ${testResults.numTotalTests}
-✅ Passed Tests: ${testResults.numPassedTests}
-❌ Failed Tests: ${testResults.numFailedTests}
-
-Pass Rate: ${testResults.numTotalTests > 0 ? ((testResults.numPassedTests / testResults.numTotalTests) * 100).toFixed(1) : 0}%
-
-`;
-
-  // Group tests by file for better organization
-  const testFiles = testResults.testResults || [];
-
-  testFiles.forEach((testFile, index) => {
-    const fileName = testFile.name
-      .replace(process.cwd(), "")
-      .replace(/^\//, "");
-    const fileStatus = testFile.status === "passed" ? "✅" : "❌";
-
-    formattedOutput += `${fileStatus} ${fileName}\n`;
-    formattedOutput += `   Tests: ${testFile.assertionResults.length} | Duration: ${testFile.endTime - testFile.startTime}ms\n`;
-
-    // Show individual test results
-    testFile.assertionResults.forEach((test) => {
-      const testStatus = test.status === "passed" ? "  ✅" : "  ❌";
-      const testName = test.fullName || test.title;
-      const duration = test.duration ? ` (${test.duration.toFixed(1)}ms)` : "";
-
-      formattedOutput += `${testStatus} ${testName}${duration}\n`;
-
-      // Show failure messages if any
-      if (
-        test.status === "failed" &&
-        test.failureMessages &&
-        test.failureMessages.length > 0
-      ) {
-        test.failureMessages.forEach((failure) => {
-          // Clean up the failure message to show just the key error
-          const errorLines = failure.split("\n");
-          const mainError =
-            errorLines.find(
-              (line) =>
-                line.includes("Error:") ||
-                line.includes("AssertionError:") ||
-                line.includes("MongoServerError:") ||
-                line.includes("CredentialsProviderError:"),
-            ) || errorLines[0];
-
-          formattedOutput += `     💥 ${mainError.trim()}\n`;
-        });
-      }
-    });
-
-    formattedOutput += `\n`;
-  });
-
-  return formattedOutput;
-}
-
-function generateHTML(data) {
-  return `
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>FG-CW-Backend Integration Test Report</title>
-    <style>
-        * { margin: 0; padding: 0; box-sizing: border-box; }
-        
-        body { 
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-            line-height: 1.6; 
-            color: #333;
-            background: #f8fafc;
-        }
-        
-        .header {
-            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-            color: white;
-            padding: 2rem 0;
-            text-align: center;
-        }
-        
-        .container { 
-            max-width: 1200px; 
-            margin: 0 auto; 
-            padding: 0 1rem; 
-        }
-        
-        .stats {
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-            gap: 1rem;
-            margin: 2rem 0;
-        }
-        
-        .stat-card {
-            background: white;
-            padding: 1.5rem;
-            border-radius: 8px;
-            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-            text-align: center;
-        }
-        
-        .stat-number {
-            font-size: 2.5rem;
-            font-weight: bold;
-            margin-bottom: 0.5rem;
-        }
-        
-        .passed { color: #10b981; }
-        .failed { color: #ef4444; }
-        .total { color: #6366f1; }
-        .rate { color: #8b5cf6; }
-        
-        .section {
-            background: white;
-            margin: 2rem 0;
-            border-radius: 8px;
-            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-            overflow: hidden;
-        }
-        
-        .section-header {
-            background: #f8fafc;
-            padding: 1rem 1.5rem;
-            border-bottom: 1px solid #e2e8f0;
-            font-weight: 600;
-            font-size: 1.1rem;
-        }
-        
-        .section-content {
-            padding: 1.5rem;
-        }
-        
-        .test-output {
-            background: #1a202c;
-            color: #e2e8f0;
-            padding: 1rem;
-            border-radius: 4px;
-            font-family: 'Monaco', 'Consolas', monospace;
-            font-size: 0.875rem;
-            overflow-x: auto;
-            white-space: pre-wrap;
-            max-height: 600px;
-            overflow-y: auto;
-        }
-        
-        .footer {
-            text-align: center;
-            padding: 2rem;
-            color: #64748b;
-            border-top: 1px solid #e2e8f0;
-            margin-top: 2rem;
-        }
-        
-        .badge {
-            display: inline-block;
-            padding: 0.25rem 0.75rem;
-            border-radius: 9999px;
-            font-size: 0.875rem;
-            font-weight: 500;
-        }
-        
-        .badge-success { background: #dcfce7; color: #166534; }
-        .badge-warning { background: #fef3c7; color: #92400e; }
-        .badge-danger { background: #fee2e2; color: #991b1b; }
-        
-        .highlight-box {
-            background: #f0f9ff;
-            border: 1px solid #0ea5e9;
-            border-radius: 6px;
-            padding: 1rem;
-            margin: 1rem 0;
-        }
-        
-        .highlight-box h3 {
-            color: #0ea5e9;
-            margin-bottom: 0.5rem;
-        }
-    </style>
-</head>
-<body>
-    <div class="header">
-        <div class="container">
-            <h1>🧪 FG-CW-Backend Integration Test Report</h1>
-            <p>Generated on ${new Date(data.timestamp).toLocaleString()}</p>
-        </div>
-    </div>
-    
-    <div class="container">
-        <div class="stats">
-            <div class="stat-card">
-                <div class="stat-number passed">${data.testsPassed}</div>
-                <div>Tests Passed</div>
-            </div>
-            <div class="stat-card">
-                <div class="stat-number failed">${data.testsFailed}</div>
-                <div>Tests Failed</div>
-            </div>
-            <div class="stat-card">
-                <div class="stat-number total">${data.totalTests}</div>
-                <div>Total Tests</div>
-            </div>
-            <div class="stat-card">
-                <div class="stat-number rate">${data.passRate}%</div>
-                <div>Pass Rate</div>
-            </div>
-        </div>
-        
-        <div class="section">
-            <div class="section-header">
-                📊 Test Summary
-                ${
-                  data.passRate >= 80
-                    ? '<span class="badge badge-success">Good</span>'
-                    : data.passRate >= 60
-                      ? '<span class="badge badge-warning">Needs Work</span>'
-                      : '<span class="badge badge-danger">Critical</span>'
-                }
-            </div>
-            <div class="section-content">
-                <div class="highlight-box">
-                    <h3>📈 Migration Progress</h3>
-                    <p><strong>Integration Test Migration:</strong> Successfully migrated tests from external fg-gas-case-working-integration repository to internal fg-cw-backend tests.</p>
-                    <p><strong>Benefits Achieved:</strong> Faster execution, no external dependencies, better debugging capabilities, CI/CD integration.</p>
-                    <p><strong>Defects Found:</strong> ${data.testsFailed} potential issues discovered for JIRA ticket creation.</p>
-                </div>
-                
-                <p><strong>Test Execution:</strong> Comprehensive API integration tests covering Users, Roles, Cases, Workflows, and Error Handling scenarios.</p>
-                <p><strong>Test Environment:</strong> TestContainers with isolated Docker services (MongoDB, LocalStack, Entra ID stub).</p>
-                <p><strong>Coverage:</strong> Full CRUD operations, data validation, concurrent processing, error handling, and data integrity checks.</p>
-            </div>
-        </div>
-        
-        <div class="section">
-            <div class="section-header">🔍 Detailed Test Output</div>
-            <div class="section-content">
-                <div class="test-output">${data.testOutput}</div>
-            </div>
-        </div>
-        
-        <div class="section">
-            <div class="section-header">📋 Next Steps</div>
-            <div class="section-content">
-                <h3>For Test Failures:</h3>
-                <ul style="margin-left: 2rem; margin-top: 0.5rem;">
-                    <li><strong>Document as JIRA defects</strong> - Use the defect documentation provided</li>
-                    <li><strong>Review API validation rules</strong> - Ensure documentation matches implementation</li>
-                    <li><strong>Fix error handling</strong> - Return proper HTTP responses instead of throwing exceptions</li>
-                    <li><strong>Improve data consistency</strong> - Verify database field mappings</li>
-                </ul>
-                
-                <h3 style="margin-top: 1rem;">For fg-gas-backend:</h3>
-                <ul style="margin-left: 2rem; margin-top: 0.5rem;">
-                    <li><strong>Resolve TestContainers issues</strong> - Fix HTML response problems</li>
-                    <li><strong>Apply same migration approach</strong> - Use this successful pattern</li>
-                    <li><strong>Create end-to-end tests</strong> - Test complete grant application workflows</li>
-                </ul>
-            </div>
-        </div>
-    </div>
-    
-    <div class="footer">
-        <p>🤖 Generated by FG-CW-Backend Test Suite | <a href="https://claude.ai/code" target="_blank">Powered by Claude Code</a></p>
-    </div>
-</body>
-</html>
-`;
-}
-
-// Run if called directly
-if (import.meta.url === `file://${process.argv[1]}`) {
-  generateComprehensiveReport();
-}
