+++ scripts/publish-create-new-case.js
@@ -22,35 +22,38 @@ const message = {
   time: "2025-03-28T11:30:52.000Z",
   source: "fg-gas-backend",
   specversion: "1.0",
-  type: "cloud.defra.development.fg-gas-backend.application.created",
+  type: "cloud.defra.development.fg-gas-backend.case.create",
   datacontenttype: "application/json",
   data: {
-    caseRef: "APPLICATION-REF-3",
+    caseRef: Math.random().toString(30).substring(2, 9),
     workflowCode: "frps-private-beta",
-    createdAt: "2025-03-27T10:34:52.000Z",
-    submittedAt: "2025-03-28T11:30:52.000Z",
-    identifiers: {
-      sbi: "SBI001",
-      frn: "FIRM0001",
-      crn: "CUST0001",
-      defraId: "DEFRA0001",
-    },
-    answers: {
-      scheme: "SFI",
-      agreementName: "Test application name",
-      year: 2025,
-      hasCheckedLandIsUpToDate: true,
-      actionApplications: [
-        {
-          parcelId: "9238",
-          sheetId: "SX0679",
-          code: "CSAM1",
-          appliedFor: {
-            unit: "ha",
-            quantity: 20.23,
+    status: "NEW",
+    payload: {
+      createdAt: "2025-03-27T10:34:52.000Z",
+      submittedAt: "2025-03-28T11:30:52.000Z",
+      identifiers: {
+        sbi: "SBI001",
+        frn: "FIRM0001",
+        crn: "CUST0001",
+        defraId: "DEFRA0001",
+      },
+      answers: {
+        scheme: "SFI",
+        agreementName: "Test application name",
+        year: 2025,
+        hasCheckedLandIsUpToDate: true,
+        actionApplications: [
+          {
+            parcelId: "9238",
+            sheetId: "SX0679",
+            code: "CSAM1",
+            appliedFor: {
+              unit: "ha",
+              quantity: 20.23,
+            },
           },
-        },
-      ],
+        ],
+      },
     },
   },
 };
@@ -60,7 +63,6 @@ console.log("Sending message to SQS queue:", queueUrl);
 if (process.argv.length === 3) {
   console.log("Sending sqs case for " + process.argv[2]);
   message.data.workflowCode = process.argv[2];
-  message.data.caseRef = "APPLICATION-PMF-001";
 }
 
 await sqs.send(
