--- src/cases/events/case-stage-updated.event.js
@@ -1,11 +1,0 @@
-import { CloudEvent } from "../../common/cloud-event.js";
-
-export class CaseStageUpdatedEvent extends CloudEvent {
-  constructor({ caseRef, previousStage, currentStage }) {
-    super("case.stage.updated", {
-      caseRef,
-      previousStage,
-      currentStage,
-    });
-  }
-}
