+++ src/common/config.js
@@ -113,12 +113,6 @@ export const config = convict({
       default: "eu-west-2",
       env: "AWS_REGION",
     },
-    caseStageUpdatedTopicArn: {
-      doc: "ARN of the SNS topic to publish case stage updates",
-      format: String,
-      default: "arn:aws:sns:eu-west-2:000000000000:case_stage_updated",
-      env: "CASE_STAGE_UPDATED_TOPIC_ARN",
-    },
     sns: {
       caseCreatedTopicArn: {
         doc: "ARN of the SNS topic to publish case created events",
